Fix 386 compilation error

+++ src/cipher/encrypt/sha256xor.go
@@ -222,7 +222,7 @@ func (s Sha256Xor) Decrypt(data []byte, password []byte) ([]byte, error) {
 
 	l := binary.LittleEndian.Uint32(dataLenBytes)
 
-	if buf.Len() > math.MaxUint32 {
+	if uint64(buf.Len()) > math.MaxUint32 {
 		return nil, ErrDataTooLarge
 	}
